requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
dataclasses-json==0.6.7
distro==1.9.0
duckdb==1.2.1
duckdb_engine==0.17.0
firecrawl-py==1.15.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langsmith==0.3.24
lxml==5.3.2
marshmallow==3.26.1
multidict==6.3.2
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.70.0
orjson==3.10.16
packaging==23.2
propcache==0.3.1
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.8.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
tenacity==8.2.3
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
websockets==15.0.1
wheel==0.45.1
yarl==1.19.0
zstandard==0.23.0
streamlit==1.28.0
pandas==2.2.3
pytest==8.3.5
